Update version ready to tag.
ExoPlayer logs the core library version (specifically `VERSION_SLASHY`) info level when the ExoPlayer is created.  This will let us see that it is the tivo modified version.   We can debate a more complex scheme, but I propose we always keep the base exoplayer version and add postfix with a tivo version