Update to keep `Parameters` not `TrackSelector`
The `DefaultTrackSelector` keeps a reference to the `ExoPlayerImpl` to send an invalidate message to the player thread to change track selection (of course).

Looking more closely at the code and sample implementation it is clear the `TrackSelectionParameters` are the safe thing to save across Activity events (onStop() to onRestart()).  So this change modifies the sample demo app and the `SimpleExoPlayerFactory` to work like this.