This works around bad codecs string issue.
This works aroudn the issue with support for 4k on Amino.  (https://w3-bugs.tivo.com/show_bug.cgi?id=524811)  My analysis is that the CODECS declartion (Concurrent) is wrong, and Amino may under report the level of support for HEVC Main Profile compared to other boxes but that does not say it is wrong.

This fix just ignores the issue if the profile is reported as HEVC Main on Amino

It is a hack.